Put layout related code in new files
